requirements.txt
schedule
paho-mqtt
eagle200_reader